Split ParseBindPath from config Get/Setters & add basic tests
